support multiple receivers with SPSP_ENDPOINTS_URL
The SPSP proxy makes a GET request to the configured SPSP_ENDPOINTS_URL
to retrieve the endpoint (or payment pointer) to proxy the SPSP query
to.
A balance id can also be provided with the SPSP endpoint to
automatically track the balance of receipts with this nonce submitted
to the /receipts API endpoint.